Handle failed bucket list creation
The creation of `Miner.buckets: list[Bucket]` can fail for a number of
reasons. One of them is when fetching a commitment of the old format (before this PR). Then it only contained a bucket name. Now the commitments contain more bucket configuration data.

This commit makes sure that the list only contains `Bucket`s and not any
None values. It also adds a log statement telling how many buckets were
added to the list.