(comms) Make sure we fetch correct peers
Before, in the beginning of the run when we only had one peer list for a
window, we ended up applying the latest one becuase of the faulty logic.
This make sure we return `None` if we try to fetch previous peer list
and don't have more than one peer list available on the bucket.