(validator) Use ordinal instead of mu for score
The ordinal is `mu-3*sigma` and is a measure on how confident on the
skill score we are. This change means that peers can increase their
weights two ways; either by ranking well compared to the other peers or
by being evaluated often. This makes it so it takes a bit longer for
peers joining mid run to prove themselves in the rankings.