(validator) Bin evaluation peers based on score
The idea is that top peers in the same performance bucket will be
evaluated against eachother. This is to make so we lower peers have
higher chance of gradually improving. It is also done to lower uid
pressure by having top peers be pitted against eachother. Potentially
making it more profitable to focus on fewer peers.